requirements.txt
streamlit
sentence-transformers
PyMuPDF
faiss-cpu
numpy
python-docx
groq
requests